// entidad-web/entidad-web/src/app/@data/model/detalleMaestra.ts
export interface DetalleMaestra {
  configuracionId: string;
  descripcion: string;
  descripcionCorta: string;
  estado: string;
  estadoConfiguracion: any;
  estadoRegistro: string;
  fechaCreacion: any;
  fechaModificacion: any;
  listaDetalleModalidad: any;
  listaEvaluacion: any;
  listaTipo: any;
  maeCabeceraId: number;
  maeDetalleId: number;
  orden: any;
  referencia: string;
  sigla: string;
  usuarioCreacion: any;
  usuarioModificacion: any;
  codProg: string;
}